MortEmp: 6 minutes 😎 (13.7M pixels)
MortEmp means Mortal Empire. It is the campaign in Total War Warhammer 2 that has the bigger map.

This map is 4096x3352 pixels, a total of 13,729,729 pixels.

The precise map used is warhammer_map_1_1_shadow.

This map isn't just representative of production inputs, it *is* a production input.

(Or was, until the game got updated with a new map, lolsob)