Returned BinarySize1 to actually be Size 1
I modified that query for the preliminary performance tests because that was the easiest solution with the gven architectural mess.

The specific problem was that GQ_Sandobx_Size_1_RLE needed the record made by GQ_ParseBitmap, but it wasn't in scope. So to go quickly, I just reparsed the bitmap in the former query by using GQ_TB_Binary_Size_1 as input.

So to change the binary under test, I would have needed to change it in 2 places. So I just change in one place where GQ_TB_Binary_Size_1 was pointing, lol.

Also, I took these early shortcuts in case I had performed a miracle and got a time under 2 minutes. Turns out I got 6 minutes, so now it's time to clean up & organize the code to do proper profiling & opimizations.