BROKEN: Thorough review of the ParseBitmap API
Here, "API" means "the record returned by ParseBitmap, both the field names and their value"

This change broke most other queries because they depend on the ParseBitmap API. They aren't yet fixed because the changes are still in progress.

The major changes are meant to make ParseBitmap decide how a Colour is represented during computation, and provide means to decompose it later for final output.

Other queries are expected to work on atomic colour values, not RGB or RGBA components. This is to minimize the amount of code that needs to care about the alpha components, notably:
1) Does the image has alpha?
2) If not, do we include a full opacity value for alpha in the output?
3) If yes, do we remove it from the output regardless?
4) etc etc etc

A notable decision was that ParseBitmap provides *both* RGB and RGBA conversion, but doesn't tell you which to use. This is because what to do with the alpha column feels more like a user preference, and I decided that ParseBitmap shouldn't be concerned with user preferences.