BROKE: Removed support for FullOuter in SelfMerge
Problem was that for FullOuter, the first row in Next and last row in This were included in results.

Also, the special casing code for FullOuter in SelfMergeSort was: 1) Getting really big, and 2) Could only go smack in the middle on the function.

So I decided that for the Full Outer, I would copy-paste the function and do the special casing there.