Removed some decorative comments
In the Power Query GUI, if there is a comment above a step, a circled 'i' will mark that step, and when the cursor hovers the name, that comment will be shown in a tooltip.

The decorative comments I removed were intended to describe *several* steps, and were thus misleading as tooltips.

So I made sure that all such remaining tooltips would be actually about the step in question.