Removed files about samples generation.
We are preparing the repo for being made public to showcase our work, and this is not what we happen people to pay attention to.

Also, it's not really relevant to this repo?

And also, it makes Github say that this project is like 91% Jupyter Notebook and 9% Bash, which is Not Correct.